Fixed MSVC 6.0 warnings

+++ sf-cvs/trunk/audacity-src/src/effects/Wahwah.cpp
@@ -95,7 +95,7 @@ bool EffectWahwah::NewTrackSimpleMono()
 
    phase = startphase;
    if (mCurChannel == Track::RightChannel)
-      phase += M_PI;
+      phase += (float)M_PI;
 
    return true;
 }
